chore: split grahpql into its own file
